requirements.txt
httpx
bs4
faust-cchardet
tensorflow==2.15
pillow
flask
